Introduce support for TypeScript and include the ESLint packages. Fix up a few additional errors that have been detected.
