--- jsconfig.json
@@ -1,6 +1,0 @@
-{
-    "compilerOptions": {
-      "baseUrl": "src"
-    },
-    "include": ["src"]
-  }
