+++ package.json
@@ -22,10 +22,14 @@
     "react-scripts": "3.0.0",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
-    "typeface-roboto": "^0.0.54",
+    "typescript": "^3.4.4",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
+    "@types/jest": "^24.0.11",
+    "@types/node": "^11.13.7",
+    "@types/react": "^16.8.14",
+    "@types/react-dom": "^16.8.4",
     "babel-polyfill": "^6.26.0",
     "codecov": "^3.3.0",
     "colors": "^1.3.3",
@@ -60,20 +64,20 @@
     "test-no-cycles": "madge \"src/index.jsx\" --circular",
     "upload-coverage-report": "./node_modules/.bin/codecov",
     "eject": "react-scripts eject",
-    "prettier": "prettier --write \"src/**/*.{js,jsx,json,css}\"",
-    "lint": "eslint \"src/**/*.{js,jsx}\" --max-warnings 0",
-    "lint-fix": "eslint \"src/**/*.{js,jsx}\" --fix --max-warnings 0 --config \".eslintrc-fixonly.js\"",
+    "prettier": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,css}\"",
+    "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\" --max-warnings 0",
+    "lint-fix": "eslint \"src/**/*.{js,jsx,ts,tsx}\" --fix --max-warnings 0 --config \".eslintrc-fixonly.js\"",
     "deploy-firestore": "firebase deploy --only firestore",
     "setup": "node utils/generate-env.js",
     "add-test-data": "node utils/generate-test-data.js",
     "cypress:open": "cypress open"
   },
   "lint-staged": {
-    "src/**/*.{js,jsx,json,css}": [
+    "src/**/*.{js,jsx,ts,tsx,json,css}": [
       "prettier --write",
       "git add"
     ],
-    "src/**/*.{js,jsx}": [
+    "src/**/*.{js,jsx,ts,tsx}": [
       "eslint --fix --max-warnings 0 --config \".eslintrc-fixonly.js\"",
       "git add"
     ]
